refactor(source): flatten parser/raster subdir and fix function names (stage 0.4)
- parser/raster/rgb.ts -> parser/rgb.ts, 函数名 rasterRgb -> rgb
- parser/raster/ndi.ts -> parser/ndi.ts, 函数名 rasterRgb -> ndi
- parser/rasterRgb.ts (注册名 rasterRgb) 保持不变, 与 rgb/ndi 功能各异
- 三个 parser 非重复代码, 本次仅命名/位置整理
- 顺带回填阶段 0.3 commit SHA (a41999c) 到 PROGRESS

git mv 保留历史, default export 函数名改动对导入方无影响,
27/27 单测通过, tsc source/src 自身 0 错误。
