+++ packages/source/src/index.ts
@@ -7,11 +7,11 @@ import image from './parser/image';
 import json from './parser/json';
 import jsonTile from './parser/jsonTile';
 import mapboxVectorTile from './parser/mvt';
+import ndi from './parser/ndi';
 import raster from './parser/raster';
 import rasterTile, { rasterDataTypes } from './parser/raster-tile';
-import ndi from './parser/raster/ndi';
-import rgb from './parser/raster/rgb';
 import rasterRgb from './parser/rasterRgb';
+import rgb from './parser/rgb';
 import testTile from './parser/testTile';
 import { cluster } from './transform/cluster';
 import { filter } from './transform/filter';
